security con permisos

+++ src/main/java/com/owo/TP_prg3/Front/Menu/MenuEntidades.java
@@ -85,7 +85,7 @@ public class MenuEntidades {
 
     private void buscarPorId() throws IOException, InterruptedException {
         System.out.print("Ingrese el ID de la entidad: ");
-        Integer id = Escaner.enteroValido(scanner);
+        Long id = Long.valueOf(Escaner.enteroValido(scanner));
 
         Optional<Object> result = HandlerResponse.handleResponse(
                 HttpService.realizarPeticion("GET", API_URL + "/" + id, authHeader, null),
